+++ js/data/setdex_ynk.js
@@ -33,15 +33,15 @@ var SETDEX_YNK = {
 	},
 	"Assist Komachi": {
 		//Below Contribution(s) by Damage
-		"Utility": {"ability": "Grace of Water", "level": 50, "pp": {"hp": 64, "fd": 64, "sp": 2}, "moves": ["First Aid", "Will-o'-Wisp", "Waterfall Drop", "Specter’s Waltz"], "mark": "Blue", "item": "Anti-Dark Charm"},
+		"Utility": {"ability": "Grace of Water", "level": 50, "pp": {"hp": 64, "fd": 64, "sp": 2}, "moves": ["First Aid", "Will-o'-Wisp", "Waterfall Drop", "Specter's Waltz"], "mark": "Blue", "item": "Anti-Dark Charm"},
 	},
 	"Assist Magan": {
 		//Below Contribution(s) by Damage
-		"Suicude Lead": {"ability": "Curiosity", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "Field Protect", "Field Barrier", "Puppet’s Grudge"], "mark": "Blue"},
+		"Suicude Lead": {"ability": "Curiosity", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "Field Protect", "Field Barrier", "Puppet's Grudge"], "mark": "Blue"},
 	},
 	"Assist Satori": {
 		//Below Contribution(s) by Damage
-		"Please Free Custap": {"ability": "Shadow Stitch", "level": 50, "pp": {"hp": 64, "sd": 64, "sp": 2}, "moves": ["Cross Counter", "Moon’s Reflection", "Encourage", "Lucky Rainbow"], "mark": "White", "item": "Life Charm"},
+		"Please Free Custap": {"ability": "Shadow Stitch", "level": 50, "pp": {"hp": 64, "sd": 64, "sp": 2}, "moves": ["Cross Counter", "Moon's Reflection", "Encourage", "Lucky Rainbow"], "mark": "White", "item": "Life Charm"},
 	},
 	"Assist Tewi": {
 		//Below Contribution(s) by Damage
@@ -50,7 +50,7 @@ var SETDEX_YNK = {
 	},
 	"Assist Udonge": {
 		//Below Contribution(s) by Damage
-		"Utility": {"ability": "Inverse Toxin", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["First Aid", "Gorgon’s Eye", "St. Elmo's Fire", "Changeling"], "mark": "Blue", "item": "Wolfsbane Root"},
+		"Utility": {"ability": "Inverse Toxin", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["First Aid", "Gorgon's Eye", "St. Elmo's Fire", "Changeling"], "mark": "Blue", "item": "Wolfsbane Root"},
 	},
 	"Defense Daiyousei": {
 		//Below Contribution(s) by Damage
@@ -63,7 +63,7 @@ var SETDEX_YNK = {
 	},
 	"Defense Eiki": {
 		//Below Contribution(s) by Damage
-		"CrokiEiki": {"ability": "Unjustness", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["Dark Power", "Specter’s Waltz", "Continue", "Unconscious Mind"], "mark": "Blue", "item": "Food Rations"},
+		"CrokiEiki": {"ability": "Unjustness", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["Dark Power", "Specter's Waltz", "Continue", "Unconscious Mind"], "mark": "Blue", "item": "Food Rations"},
 	},
 	"Defense Hina": {
 		//Below Contribution(s) by Damage
@@ -144,7 +144,7 @@ var SETDEX_YNK = {
 	},
 	"Extra Doremy": {
 		//Below Contribution(s) by Damage
-		"Physical Attacker": {"ability": "Dream World", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Gorgon’s Eye", "Strenuous Stance", "Hallucination", "Hallucination"], "mark": "Green", "item": "Straw Doll"},
+		"Physical Attacker": {"ability": "Dream World", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Gorgon's Eye", "Strenuous Stance", "Hallucination", "Hallucination"], "mark": "Green", "item": "Straw Doll"},
 	},
 	"Extra Eternity": {
 		//Below Contribution(s) by Lightli
@@ -152,7 +152,7 @@ var SETDEX_YNK = {
 	},
 	"Extra Futo": {
 		//Below Contribution(s) by Damage
-		"Physical Attacker": {"ability": "Sore Loser", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Specter’s Waltz", "Heavenly Ascent", "Shooting Press", "Changeling"], "mark": "Green", "item": "Straw Doll"},
+		"Physical Attacker": {"ability": "Sore Loser", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Specter's Waltz", "Heavenly Ascent", "Shooting Press", "Changeling"], "mark": "Green", "item": "Straw Doll"},
 	},
 	"Extra Gengetsu": {
 		//Below Contribution(s) by Eevee
@@ -398,5 +398,195 @@ var SETDEX_YNK = {
 	"Speed Yumeko": {
 		//Below Contribution(s) by Barudo
 		"All-Out Attacker": {"level": 50, "pp": {"fa": 64, "sd": 2, "sp": 64}, "mark": "Green", "ability": "Salvo", "item": "Choice Belt", "moves": ["Brandish", "Shooting Arts", "Hornet's Flit", "Changeling"]},
+	},
+};
+
+var SETDEX_SOD = {
+	"Assist Chen": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Trapper": {"ability": "Adverse Wind", "level": 50, "pp": {"hp": 2, "sa": 64, "sp": 64}, "ranks": {"fa": 0}, "moves": ["Brave Song", "Encourage", "Swift Air-Raid", "Waves of Earth"], "mark": "Green", "item": "Black Choker"},
+	},
+	"Assist Akyuu": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Meme Trapper - SoD": {"ability": "Frail Health", "level": 50, "pp": {"fa": 64, "sp": 64}, "moves": ["Ruinous Voice", "Black Hole", "Thorned Ivy", "Sneak Attack"], "mark": "Red", "item": "Black Choker"},
+	},
+	"Assist Cirno": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Support": {"ability": "Curiosity", "level": 50, "pp": {"hp": 64, "sd": 64, "sa": 2}, "ranks": {"fa": 0}, "moves": ["Mine Trap", "Puppet's Grudge", "Lazy Mist", "Miasma"], "mark": "White", "item": "Food Rations"},
+	},
+	"Assist Eirin": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive": {"ability": "Emergency Cure", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["First Aid", "Lazy Mist", "Smash Spin", "Poison Trap"], "mark": "Blue", "item": "Jar of Poison"},
+	},
+	"Assist Kasen": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive Pivot": {"ability": "Absorbent", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["First Aid", "Solar Eclipse", "Brave Burst", "Starving Spirit"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Assist Koakuma": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Sticky Web Lead": {"ability": "Curse Return", "level": 50, "pp": {"fd": 2, "fa": 64, "sp": 64}, "moves": ["Bind Trap", "Will-o'-Wisp", "Prank", "Changeling"], "mark": "Green", "item": "Black Choker"},
+	},
+	"Assist Komachi": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Utility": {"ability": "Grace of Water", "level": 50, "pp": {"hp": 64, "fd": 64, "sp": 2}, "moves": ["First Aid", "Will-o'-Wisp", "Waterfall Drop", "Specter's Waltz"], "mark": "Blue", "item": "Anti-Dark Charm"},
+	},
+	"Assist Magan": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Suicude Lead": {"ability": "Curiosity", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "Field Protect", "Field Barrier", "Puppet's Grudge"], "mark": "Blue"},
+	},
+	"Assist Satori": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Please Free Custap": {"ability": "Shadow Stitch", "level": 50, "pp": {"hp": 64, "sd": 64, "sp": 2}, "moves": ["Cross Counter", "Moon's Reflection", "Encourage", "Lucky Rainbow"], "mark": "White", "item": "Life Charm"},
+	},
+	"Assist Tewi": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Support": {"ability": "Lucky", "level": 50, "pp": {"hp": 2, "fd": 64, "sp": 64}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "First Aid", "Lightning Speed", "Destruction"], "mark": "Green", "item": "Food Rations"},
+		"Suicide Lead": {"ability": "Quick Eye", "level": 50, "pp": {"hp": 2, "sa": 64, "sp": 64}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "Bind Trap", "Upbeat", "Destruction"], "mark": "Green", "item": "Black Choker"},
+	},
+	"Assist Udonge": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Utility": {"ability": "Inverse Toxin", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["First Aid", "Gorgon's Eye", "St. Elmo's Fire", "Changeling"], "mark": "Blue", "item": "Wolfsbane Root"},
+	},
+	"Defense Daiyousei": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive Pivot": {"ability": "Glamorous", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Bewitching Pollen", "Grace of Mana", "Pollen Mist", "Changeling"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Defense Doremy": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"Focus Wall": {"level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "mark": "Blue", "ability": "Dark Clothing", "item": "Counter Bit", "moves": ["Grace of Mana", "Changeling", "Poltergeist", "Ruinous Voice"]},
+	},
+	"Defense Eiki": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"CrokiEiki": {"ability": "Unjustness", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["Dark Power", "Specter's Waltz", "Continue", "Unconscious Mind"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Defense Hina": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Physically Defensive": {"ability": "Air Cushion", "level": 50, "pp": {"hp": 64, "fd": 64, "sp": 2}, "ranks": {"fa": 0}, "moves": ["Poison Bomb", "Poison Trap", "Will-o'-Wisp", "Upbeat"], "mark": "Blue", "item": "Jar of Poison"},
+	},
+	"Defense Koakuma": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive": {"ability": "Healing Power", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Power Spot", "Supernatural Border", "Impulse", "Changeling"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Defense Medicine": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive Pivot": {"ability": "Karmic Retribution", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Bout Drunkard", "Grace of Mana", "Poison Stream", "Power Drain"], "mark": "Blue", "item": "Jar of Poison"},
+	},
+	"Defense Ringo": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive": {"ability": "In Sync", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "First Aid", "Voice of Famine", "Atomic Energy"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Defense Orange": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive": {"ability": "Emergency Cure", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Aikido Arts", "Miasma", "First Aid", "Forest Therapy"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Defense Sara": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Mixed Wall": {"ability": "Choleric", "level": 50, "pp": {"hp": 64, "fd": 20, "sd": 46}, "moves": ["First Aid", "Stealth Trap", "Shooting Press", "Miasma"], "mark": "White", "item": "Food Rations"},
+	},
+	"Defense Satori": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Utility": {"ability": "Curse Return", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["Decrescendo", "St. Elmo's Fire", "Revolving Illusions", "Miasma"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Defense Tokiko": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive Utility": {"ability": "Deploy Smoke", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["Stealth Trap", "Perch", "Changeling", "Air Stamp"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Extra Akyuu": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Suicide Lead": {"ability": "Curiosity", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["A Tempo", "Recall Nightmare", "Eternal Record", "Forest Record"], "mark": "Blue", "item": "Dispel Charm"},
+	},
+	"Extra Aya": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"Pivot": {"level": 50, "pp": {"hp": 64, "fd": 2, "sd": 64}, "mark": "Blue", "ability": "Intuition", "item": "Reflect Bit", "moves": ["Lightning Speed", "Spinning Air", "Perch", "Backdraft"]},
+	},
+	"Extra Chen": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Belly Drum": {"ability": "Known Limits", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Inner Power", "Swallow Cut", "Shooting Press", "Flying Frenzy"], "mark": "Red", "item": "Life Charm"},
+	},
+	"Extra Cirno": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Utility": {"ability": "Known Limits", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Stealth Trap", "Grace of Mana", "Lazy Mist", "Lightning Speed"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Extra Doremy": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Physical Attacker": {"ability": "Dream World", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Gorgon's Eye", "Strenuous Stance", "Hallucination", "Hallucination"], "mark": "Green", "item": "Straw Doll"},
+	},
+	"Extra Futo": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Physical Attacker": {"ability": "Sore Loser", "level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "moves": ["Specter's Waltz", "Heavenly Ascent", "Shooting Press", "Changeling"], "mark": "Green", "item": "Straw Doll"},
+	},
+	"Extra Hatate": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Unburden": {"ability": "Nimble", "level": 50, "pp": {"sa": 64, "sp": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Raid", "Spinning Air", "Impulse", "Brave Burst"], "mark": "Green", "item": "Diamond"},
+	},
+	"Extra Hina": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"Hairpin Pivot": {"level": 50, "pp": {"hp": 64, "fa": 64, "fd": 2}, "mark": "Red", "ability": "Air Cushion", "item": "Silver Hairpin", "moves": ["Holy Flare", "Poison Stream", "Changeling", "Shadow Bomb"]},
+		"Drain Seed": {"level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "mark": "Blue", "ability": "Air Cushion", "item": "Jar of Poison", "moves": ["Drain Seed", "Supernatural Border", "Holy Flare", "Poison Stream"]},
+	},
+	"Extra Kikuri": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive": {"ability": "In Sync", "level": 50, "pp": {"hp": 11, "fd": 64, "sd": 56}, "moves": ["Power Spot", "Supernatural Border", "Miasma", "Revolving Illusions"], "mark": "Blue", "item": "Food Rations"},
+	},
+	"Extra Medicine": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"Lunar Lockdown": {"level": 50, "pp": {"hp": 12, "sa": 54, "sp": 64}, "mark": "Green", "ability": "Poison Labyrinth", "item": "Choice Belt", "moves": ["Shark Trade", "Moon's Protection", "Impulse", "Continue"]},
+		"SleepTrap": {"level": 50, "pp": {"fd": 2, "sa": 64, "sp": 64}, "mark": "Green", "ability": "Poison Labyrinth", "item": "Jar of Poison", "moves": ["Bewitching Pollen", "Magic Barrier", "Poison Bomb", "Impulse"]},
+		"Revenge Killer": {"level": 50, "pp": {"fd": 2, "sa": 64, "sp": 64}, "mark": "Green", "ability": "Poison Labyrinth", "item": "Choice Earrings", "moves": ["Impulse", "Sweet Desperado", "Shark Trade", "Pollen Mist"]},
+		"Poison Trap Lead": {"level": 50, "pp": {"fd": 2, "sa": 64, "sp": 64}, "mark": "Green", "ability": "Poison Labyrinth", "item": "Black Choker", "moves": ["Poison Trap", "Poison Bomb", "Impulse", "Upbeat"]},
+	},
+	"Extra Miko": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Nasty Plot Sweeper": {"ability": "Intuition", "level": 50, "pp": {"sa": 64, "sp": 64, "sd": 2}, "ranks": {"fa": 0}, "moves": ["Brave Song", "Phase Inversion", "Backdraft", "Ultra High Tone"], "mark": "Green", "item": "Straw Doll"},
+	},
+	"Extra Momiji": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive Utility": {"ability": "Counter Stance", "level": 50, "pp": {"hp": 64, "fd": 23, "sd": 43}, "moves": ["Stealth Trap", "Drain Seed", "Power Drain", "Impact Rebellion"], "mark": "White", "item": "Food Rations"},
+	},
+	"Extra Remilia": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Fog Breaker": {"ability": "Fog Traveler", "level": 50, "pp": {"sa": 64, "sp": 64, "hp": 2}, "moves": ["Frozen Prison", "Poison Bomb", "Eye of Laplace", "Perch"], "mark": "Green", "item": "Straw Doll"},
+	},
+	"Extra Renko": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Sticky Web Setter": {"ability": "Invalidate", "level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "moves": ["Stealth Trap", "Bind Trap", "Encourage", "Miasma"], "mark": "Blue", "item": "Black Choker"},
+	},
+	"Extra Sagume": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"All-Out Attacker": {"level": 50, "pp": {"hp": 64, "fa": 64, "fd": 2}, "mark": "Red", "ability": "Reverse Function", "item": "Silver Hairpin", "moves": ["Distortion Bomb", "Trickster", "High Tone Crush", "Clearing Mist"]},
+	},
+	"Extra Suwako": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Defensive": {"level": 50, "pp": {"hp": 64, "fd": 64, "sd": 2}, "ability": "Negative Aura", "moves": ["First Aid", "Acid Tears", "Poison Stream", "Distortion Bomb"], "mark": "Blue", "item": "Jar of Poison"},
+	},
+	"Extra Toyohime": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Silver Hairpin Pivot": {"level": 50, "pp": {"hp": 64, "sd": 64, "fd": 2}, "ability": "Healing Power", "moves": ["Energy Absorb", "Moonbow", "Smash Spin", "Smash Spin"], "mark": "White", "item": "Silver Hairpin"},
+	},
+	"Extra Wakasagi": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"Genbu Breaker": {"level": 50, "pp": {"hp": 64, "sa": 64, "sd": 2}, "mark": "Black", "ranks": {"sp": 0}, "ability": "Gentei's Water", "item": "Dispel Charm", "moves": ["Terrain [Genbu]", "Earthly Blessing", "Energy Absorb", "Brave Burst"]},
+	},
+	"Power Aya": {
+		//Below Contribution(s) by Damage from the TPDP Discord
+		"All-Out Attacker": {"level": 50, "pp": {"fa": 64, "sp": 64, "hp": 2}, "mark": "Green", "ability": "Salvo", "moves": ["Twister", "Shooting Arts", "Godstone Frenzy", "Lightning Speed"]},
+	},
+	"Power Konngara": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"Hairpin": {"level": 50, "pp": {"hp": 64, "fa": 64, "fd": 2}, "mark": "Red", "ability": "Benefit of Fire", "item": "Silver Hairpin", "moves": ["Heavenly Ascent", "Samurai Edge", "Waterfall Drop", "Strike Shot"]},
+		"Strenuous Stance": {"level": 50, "pp": {"fa": 64, "fd": 2, "sp": 64}, "mark": "Red", "ability": "Benefit of Fire", "item": "Black Choker", "moves": ["Strenuous Stance", "Heavenly Ascent", "Samurai Edge", "Waterfall Drop"]},
+	},
+	"Power Sakuya": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"All-Out Attacker": {"level": 50, "pp": {"hp": 64, "fa": 64, "fd": 2}, "mark": "Red", "ability": "Strategist", "item": "Golden Hairpin", "moves": ["Overtake Strike", "Brandish", "Dancing Sword", "Changeling"]},
+	},
+	"Speed Junko": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"All-Out Attacker": {"level": 50, "pp": {"fd": 2, "sa": 64, "sp": 64}, "mark": "Green", "ability": "Hate Incarnate", "item": "Choice Earrings", "moves": ["Dark Sweets", "Moonbow", "Impulse", "Mysterious Wave"]},
+	},
+	"Speed Sanae": {
+		//Below Contribution(s) by MQJinx from the TPDP Discord
+		"Sharp Wind": {"level": 50, "pp": {"fd": 2, "sa": 64, "sp": 64}, "mark": "Green", "ability": "Common Senseless", "item": "Black Choker", "moves": ["Sharp Wind", "Waves of Earth", "Macroburst", "Pollen Mist"]},
+		"Choice Belt": {"level": 50, "pp": {"fd": 2, "sa": 64, "sp": 64}, "mark": "Green", "ability": "Common Senseless", "item": "Black Choker", "moves": ["Macroburst", "Earthen Feast", "Pollen Mist", "Unknown Flare"]},
 	},
 };